requirements.txt
PyPDF2
python-docx 
tox==3.25.1
chromadb
sentence_transformers
python-dotenv
-e .